Workaround quadratic memory BIO read in BoringSSL
If one were to read a 1MB BIO_s_mem byte-by-byte, BoringSSL's
accidentally-quadratic implementation ends up memcpying
512GB of data about.  `BIO_new_mem_buf` doesn't (even though it is
backed by the same code -- why not do it in the same way?!)

This is responsible for destroying BoringSSL's read performance
in this benchmark.

OpenSSL is not affected because they fixed this in 2016, though
their manpage still documents the performance issue.